[feat] all id and price done

+++ features/entertainment/digital_deals/src/main/java/com/tokopedia/digital_deals/view/model/ProductItem.java
@@ -30,11 +30,11 @@ public class ProductItem extends BaseItem implements Parcelable {
     private String longRichDesc;
     @SerializedName("mrp")
     @Expose
-    private int mrp;
+    private long mrp;
     @SuppressLint("Invalid Data Type")
     @SerializedName("sales_price")
     @Expose
-    private int salesPrice;
+    private long salesPrice;
     @SerializedName("quantity")
     @Expose
     private int quantity;
@@ -95,7 +95,7 @@ public class ProductItem extends BaseItem implements Parcelable {
     @SuppressLint("Invalid Data Type")
     @SerializedName("id")
     @Expose
-    private int id;
+    private long id;
     @SerializedName("is_searchable")
     @Expose
     private int isSearchable;
@@ -167,8 +167,8 @@ public class ProductItem extends BaseItem implements Parcelable {
         imageWeb = in.readString();
         thumbnailWeb = in.readString();
         longRichDesc = in.readString();
-        mrp = in.readInt();
-        salesPrice = in.readInt();
+        mrp = in.readLong();
+        salesPrice = in.readLong();
         quantity = in.readInt();
         soldQuantity = in.readInt();
         sellRate = in.readInt();
@@ -275,19 +275,19 @@ public class ProductItem extends BaseItem implements Parcelable {
         this.longRichDesc = longRichDesc;
     }
 
-    public int getMrp() {
+    public long getMrp() {
         return mrp;
     }
 
-    public void setMrp(int mrp) {
+    public void setMrp(long mrp) {
         this.mrp = mrp;
     }
 
-    public int getSalesPrice() {
+    public long getSalesPrice() {
         return salesPrice;
     }
 
-    public void setSalesPrice(int salesPrice) {
+    public void setSalesPrice(long salesPrice) {
         this.salesPrice = salesPrice;
     }
 
@@ -484,11 +484,11 @@ public class ProductItem extends BaseItem implements Parcelable {
         isLiked = liked;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -508,8 +508,8 @@ public class ProductItem extends BaseItem implements Parcelable {
         dest.writeString(imageWeb);
         dest.writeString(thumbnailWeb);
         dest.writeString(longRichDesc);
-        dest.writeInt(mrp);
-        dest.writeInt(salesPrice);
+        dest.writeLong(mrp);
+        dest.writeLong(salesPrice);
         dest.writeInt(quantity);
         dest.writeInt(soldQuantity);
         dest.writeInt(sellRate);
@@ -526,7 +526,7 @@ public class ProductItem extends BaseItem implements Parcelable {
         dest.writeString(cityName);
         dest.writeInt(rating);
         dest.writeInt(likes);
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeInt(isSearchable);
         dest.writeString(categoryId);
         dest.writeString(providerId);
